app: image and video like markdown rendered as page
